requirements.txt
flask>=2.3.3
flask-sqlalchemy>=3.0.5
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.35
python-decouple>=3.8
waitress>=3.0.0
python-dotenv>=1.0.1
gunicorn>=23.0.0
twilio>=9.3.3
google-cloud-texttospeech>=2.17.2
google-auth>=2.35.0
openai>=1.51.0
langgraph>=0.2.34
langsmith>=0.1.125
langchain-core>=0.3.7
pandas>=2.2.3
numpy>=2.1.2
faiss-cpu>=1.9.0
alembic>=1.13.3